Optimize image copying loop after capturing an image
This improves the runtime of this loop from about 10ms to less than 1ms
on my computer.

When repeatedly obtaining images through the API in a loop, this image
copying does not seem to be in the critical path (probably because the GPU
already starts rendering the next frame while the copy is going on), but
it seems to be some low-hanging fruit nonetheless.